+++ packages/react-native/CHANGELOG.md
@@ -1,1 +1,3 @@
+## Next
 
+* Re-export `<Curious>`, not `curious()`.